+++ metadata.rb
@@ -2,9 +2,9 @@ name 'incident_bot'
 maintainer 'Hearst Automation Team'
 maintainer_email 'atat@hearst.com'
 license 'All rights reserved'
-description 'Installs/Configures incident_bot'
+description 'Installs/Configures the Hearst Incident Bot'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.1.0'
+version '1.0.0'
 
 supports 'ubuntu', '>= 10.04'
 
@@ -16,3 +16,5 @@ depends 'nginx'
 depends 'nodejs'
 depends 'apt'
 depends 'git'
+depends 'chef-client'
+depends 'citadel'