+++ recipes/_supervisor.rb
@@ -1,5 +1,10 @@
 include_recipe 'supervisor'
 
+node.default['incident_bot']['config']['HUBOT_SLACK_TOKEN'] = citadel['slack/api_key']
+node.default['incident_bot']['config']['HUBOT_PAGERDUTY_USER_ID'] = citadel['pagerduty/user_id']
+node.default['incident_bot']['config']['HUBOT_PAGERDUTY_SERVICE_API_KEY'] = citadel['pagerduty/service_key']
+node.default['incident_bot']['config']['HUBOT_PAGERDUTY_USER_ID'] = citadel['pagerduty/user_id']
+
 hubot_name = node['incident_bot']['name']
 hubot_bin = File.join(node['incident_bot']['install_dir'], 'node_modules', '.bin', 'hubot')
 hubot_adapter = node['incident_bot']['adapter']