+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# 1.0.0
+First function release of the hearst incident bot (Only Chef-Zero capable at the moment)
+* Citadel Enabled secrets
+* Nginx HTTPS Proxy to Hubot Webhooks
+* Letsencrypt Certificates and Auto-Renewal
+* Redis Brain persistance through S3FS (S3FS configured by CFN for the time being)
+
 # 0.1.0
 Initial release of incident-bot